expose range

+++ include/tvm/expr.h
@@ -89,7 +89,7 @@ class Var : public Expr {
 };
 
 Expr IntConstant(int64_t value);
-Expr FloatConstant(int64_t value);
+Expr FloatConstant(double value);
 
 /*! \brief base of expression node */
 class ExprNode : public Node {
